refactor(retrieval): split gist.rs into submodules, fix benchmark imports
- Split gist.rs (997 LOC) into 6 files under 500 LOC limit:
  - mod.rs (75 LOC): re-exports and documentation
  - config.rs (97 LOC): RerankConfig struct
  - types.rs (52 LOC): EpisodeGist struct
  - extractor.rs (271 LOC): GistExtractor impl
  - reranker.rs (311 LOC): HierarchicalReranker
  - tests.rs (222 LOC): all 16 tests

- Fix benchmark crate name imports (memory_benches → do_memory_benches)
- Add missing checkpoints field to domain_cache_benchmark Episode
- Update cli_benchmarks to use do_memory_cli::test_utils
- Update CHANGELOG and STATUS for v0.1.31 preparation
