+++ benches/concurrent_operations.rs
@@ -6,13 +6,13 @@
 #![allow(clippy::excessive_nesting)]
 #![allow(deprecated)]
 use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
-use do_memory_core::types::{TaskOutcome, TaskType};
-use futures::future::join_all;
-use memory_benches::TokioExecutor;
-use memory_benches::benchmark_helpers::{
+use do_memory_benches::TokioExecutor;
+use do_memory_benches::benchmark_helpers::{
     create_benchmark_context, generate_episode_description, generate_execution_steps,
     setup_temp_memory,
 };
+use do_memory_core::types::{TaskOutcome, TaskType};
+use futures::future::join_all;
 use rand::{RngExt, SeedableRng};
 use std::sync::Arc;
 use tokio::sync::Semaphore;
