+++ benches/genesis_benchmark.rs
@@ -10,16 +10,16 @@
 #![allow(clippy::excessive_nesting)]
 #![allow(deprecated)]
 use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
+use do_memory_benches::TokioExecutor;
+use do_memory_benches::benchmark_helpers::{
+    create_benchmark_context, generate_episode_description, generate_execution_steps,
+    setup_temp_memory,
+};
 use do_memory_core::{
     Episode,
     episodic::{CapacityManager, EvictionPolicy},
     semantic::SemanticSummarizer,
     types::{TaskOutcome, TaskType},
-};
-use memory_benches::TokioExecutor;
-use memory_benches::benchmark_helpers::{
-    create_benchmark_context, generate_episode_description, generate_execution_steps,
-    setup_temp_memory,
 };
 use std::hint::black_box;
 use std::time::Instant;
