+++ benches/soak_tests.rs
@@ -12,11 +12,11 @@
 #![allow(clippy::excessive_nesting)]
 #![allow(deprecated)]
 use criterion::{criterion_group, criterion_main, Criterion};
-use memory_benches::benchmark_helpers::{
+use do_memory_benches::benchmark_helpers::{
     create_benchmark_context, generate_episode_description, generate_execution_steps,
     setup_temp_memory,
 };
-use memory_benches::TokioExecutor;
+use do_memory_benches::TokioExecutor;
 use do_memory_core::types::{TaskOutcome, TaskType};
 use std::sync::atomic::{AtomicBool, AtomicU64, AtomicUsize, Ordering};
 use std::sync::Arc;
