+++ benches/storage_operations.rs
@@ -3,13 +3,13 @@
 #![allow(clippy::excessive_nesting)]
 #![allow(deprecated)]
 use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
-use do_memory_core::types::{TaskOutcome, TaskType};
-use do_memory_storage_redb::RedbStorage;
-use memory_benches::TokioExecutor;
-use memory_benches::benchmark_helpers::{
+use do_memory_benches::TokioExecutor;
+use do_memory_benches::benchmark_helpers::{
     create_benchmark_context, generate_episode_description, generate_execution_steps,
     setup_temp_memory,
 };
+use do_memory_core::types::{TaskOutcome, TaskType};
+use do_memory_storage_redb::RedbStorage;
 use std::collections::HashMap;
 use std::hint::black_box;
 use uuid::Uuid;
